+++ hash_checks/horizon_checks.py
@@ -2,7 +2,7 @@ from pathlib import Path
 from datetime import datetime
 from zoneinfo import ZoneInfo
 
-from starplot import styles, HorizonPlot, _, Observer, Star
+from starplot import styles, HorizonPlot, _, Observer, Star, CollisionHandler
 
 HERE = Path(__file__).resolve().parent
 DATA_PATH = HERE / "data"
@@ -14,7 +14,7 @@ STYLE = styles.PlotStyle().extend(
 
 RESOLUTION = 4096
 
-AUTO_ADJUST_SETTINGS = {"seed": 1}
+HANDLER = CollisionHandler(seed=1, allow_constellation_line_collisions=True)
 
 TZ_PT = ZoneInfo("US/Pacific")
 
@@ -39,9 +39,10 @@ def _horizon():
     p.constellation_borders()
     p.milky_way()
     p.stars(where=[_.magnitude < 5])
+    p.messier(where_true_size=[_.size > 1])
     p.ecliptic()
     p.horizon()
-    p.constellation_labels(auto_adjust_settings=AUTO_ADJUST_SETTINGS)
+    p.constellation_labels(collision_handler=HANDLER)
     p.gridlines()
     return p
 
@@ -73,9 +74,10 @@ def check_horizon_north_celestial_pole():
     p.constellation_borders()
     p.milky_way()
     p.stars(where=[_.magnitude < 5])
+    p.messier(where_true_size=[_.size > 1])
     p.ecliptic()
     p.horizon()
-    p.constellation_labels(auto_adjust_settings=AUTO_ADJUST_SETTINGS)
+    p.constellation_labels(collision_handler=HANDLER)
     p.gridlines()
 
     polaris = Star.get(name="Polaris")
@@ -110,7 +112,7 @@ def check_horizon_gradient_background():
     p.stars(where=[_.magnitude < 5])
     p.ecliptic()
     p.horizon()
-    p.constellation_labels(auto_adjust_settings=AUTO_ADJUST_SETTINGS)
+    p.constellation_labels(collision_handler=HANDLER)
     p.gridlines()
 
     filename = DATA_PATH / "horizon-gradient-background.png"