+++ hash_checks/zenith_checks.py
@@ -2,7 +2,14 @@ from pathlib import Path
 from datetime import datetime
 from zoneinfo import ZoneInfo
 
-from starplot import ZenithPlot, Observer, styles, override_settings, _
+from starplot import (
+    ZenithPlot,
+    Observer,
+    styles,
+    override_settings,
+    CollisionHandler,
+    _,
+)
 
 HERE = Path(__file__).resolve().parent
 DATA_PATH = HERE / "data"
@@ -14,6 +21,8 @@ STYLE = styles.PlotStyle().extend(
 JUNE_2023 = datetime(2023, 6, 20, 21, 0, 0, 0, tzinfo=ZoneInfo("US/Pacific"))
 
 RESOLUTION = 3800
+
+HANDLER = CollisionHandler(seed=1, allow_constellation_line_collisions=True)
 
 
 def _zenith():
@@ -34,7 +43,7 @@ def _zenith():
     p.ecliptic(style__line__width=8)
     p.celestial_equator(style__line__width=8)
     p.legend(style__location=styles.LegendLocationEnum.INSIDE_BOTTOM_RIGHT)
-    p.constellation_labels()
+    p.constellation_labels(collision_handler=HANDLER)
     return p
 
 
@@ -66,7 +75,7 @@ def check_zenith_gradient():
     p.stars(where=[_.magnitude < 4.6], where_labels=[_.magnitude < 3])
     p.ecliptic(style__line__width=8)
     p.celestial_equator(style__line__width=8)
-    p.constellation_labels()
+    p.constellation_labels(collision_handler=HANDLER)
     filename = DATA_PATH / "zenith-gradient.png"
     p.export(filename)
     return filename
@@ -105,7 +114,7 @@ def check_zenith_chinese():
     p.horizon()
     p.constellations()
     p.stars(where=[_.magnitude < 4.6], where_labels=[_.magnitude < 4])
-    p.constellation_labels()
+    p.constellation_labels(collision_handler=HANDLER)
     filename = DATA_PATH / "zenith-chinese.png"
     p.export(filename)
     return filename