+++ src/starplot/data/__init__.py
@@ -1,9 +1,11 @@
+# ruff: noqa: F401,F403
+
 from pathlib import Path
 
 from skyfield.api import Loader
 
 from starplot.config import settings
-from .catalogs import Catalog  # noqa: F401
+from .catalogs import Catalog
 
 load = Loader(settings.data_path)  # used for loading ephemeris
 
@@ -16,9 +18,5 @@ INTERNAL_DATA_PATH = HERE / "library"
 
 class DataFiles:
     STAR_DESIGNATIONS = INTERNAL_DATA_PATH / "star_designations.parquet"
-
     CONSTELLATION_NAMES = INTERNAL_DATA_PATH / "constellation_names.parquet"
-
     DSO_NAMES = INTERNAL_DATA_PATH / "dso_names.parquet"
-
-    DATABASE = INTERNAL_DATA_PATH / "sky.db"