+++ src/starplot/plots/__init__.py
@@ -1,0 +1,6 @@
+# ruff: noqa: F401,F403
+
+from .map import MapPlot
+from .horizon import HorizonPlot
+from .zenith import ZenithPlot
+from .optic import OpticPlot