+++ src/starplot/plotters/__init__.py
@@ -1,7 +1,9 @@
-from .constellations import ConstellationPlotterMixin  # noqa: F401
-from .stars import StarPlotterMixin  # noqa: F401
-from .dsos import DsoPlotterMixin  # noqa: F401
-from .milkyway import MilkyWayPlotterMixin  # noqa: F401
-from .legend import LegendPlotterMixin  # noqa: F401
-from .gradients import GradientBackgroundMixin  # noqa: F401
-from .arrow import ArrowPlotterMixin  # noqa: F401
+# ruff: noqa: F401,F403
+
+from .constellations import ConstellationPlotterMixin
+from .stars import StarPlotterMixin
+from .dsos import DsoPlotterMixin
+from .milkyway import MilkyWayPlotterMixin
+from .legend import LegendPlotterMixin
+from .gradients import GradientBackgroundMixin
+from .arrow import ArrowPlotterMixin