+++ src/starplot/plotters/constellations.py
@@ -10,8 +10,13 @@ from matplotlib.collections import LineCollection
 from ibis import _
 
 from starplot.coordinates import CoordinateSystem
-from starplot.data import constellations as condata
-from starplot.data.catalogs import Catalog, CONSTELLATIONS_IAU, BIG_SKY_MAG11
+from starplot.data import db, constellations as condata
+from starplot.data.catalogs import (
+    Catalog,
+    CONSTELLATIONS_IAU,
+    CONSTELLATION_BORDERS,
+    BIG_SKY_MAG11,
+)
 from starplot.data.stars import load as load_stars
 from starplot.models import Star, Constellation
 from starplot.models.constellation import from_tuple
@@ -27,17 +32,7 @@ from starplot.styles import LineStyle, LabelStyle
 from starplot.styles.helpers import use_style
 from starplot.utils import points_on_line
 from starplot.geometry import is_wrapped_polygon
-
-DEFAULT_AUTO_ADJUST_SETTINGS = {
-    "avoid_constellation_lines": False,
-    "point_generation_max_iterations": 10,
-    "distance_step_size": 2,
-    "max_distance": 3_000,
-    "label_padding": 6,
-    "buffer": 0.3,
-    "seed": None,
-}
-"""Default settings for auto-adjusting constellation labels"""
+from starplot.plotters.text import CollisionHandler
 
 GEODETIC_PROJECTIONS = (
     Equidistant,
@@ -74,6 +69,7 @@ class ConstellationPlotterMixin:
             filters=[_.hip.isin(hips)],
         )
         df = results.to_pandas()
+        df["ra_hours"], df["dec_degrees"] = (df.ra / 15, df.dec)
         df = self._prepare_star_coords(df, limit_by_altaz=False)
 
         return {star.hip: (star.x, star.y) for star in df.itertuples()}
@@ -95,6 +91,7 @@ class ConstellationPlotterMixin:
             style: Styling of the constellations. If None, then the plot's style (specified when creating the plot) will be used
             where: A list of expressions that determine which constellations to plot. See [Selecting Objects](/reference-selecting-objects/) for details.
             sql: SQL query for selecting constellations (table name is `_`). This query will be applied _after_ any filters in the `where` kwarg.
+            catalog: The catalog of constellations to use -- see [catalogs overview](/data/overview/) for details
         """
         self.logger.debug("Plotting constellation lines...")
 
@@ -126,7 +123,7 @@ class ConstellationPlotterMixin:
             inbounds = False
 
             for s1_hip, s2_hip in hiplines:
-                if not constars.get(s2_hip):
+                if not constars.get(s1_hip) or not constars.get(s2_hip):
                     continue
                 s1_ra, s1_dec = constars.get(s1_hip)
                 s2_ra, s2_dec = constars.get(s2_hip)
@@ -206,15 +203,23 @@ class ConstellationPlotterMixin:
 
     @profile
     @use_style(LineStyle, "constellation_borders")
-    def constellation_borders(self, style: LineStyle = None):
+    def constellation_borders(
+        self, style: LineStyle = None, catalog: Catalog = CONSTELLATION_BORDERS
+    ):
         """Plots the constellation borders
 
         Args:
             style: Styling of the constellation borders. If None, then the plot's style (specified when creating the plot) will be used
+            catalog: Catalog to use for constellation borders
         """
+        con = db.connect()
+        borders = catalog._load(connection=con, table_name="constellation_borders")
+        borders = borders.mutate(
+            geometry=_.geometry.cast("geometry"),  # cast WKB to geometry type
+        )
+
         extent = self._extent_mask()
-        results = condata.load_borders(extent=extent)
-        borders_df = results.to_pandas()
+        borders_df = borders.filter(_.geometry.intersects(extent)).to_pandas()
 
         if borders_df.empty:
             return
@@ -248,7 +253,29 @@ class ConstellationPlotterMixin:
         )
         self.ax.add_collection(line_collection)
 
-    def _constellation_labels_auto(self, style, label_fn, settings):
+    @profile
+    @use_style(LabelStyle, "constellation_labels")
+    def constellation_labels(
+        self,
+        style: LabelStyle = None,
+        label_fn: Callable[[Constellation], str] = Constellation.get_label,
+        collision_handler: CollisionHandler = None,
+    ):
+        """
+        Plots constellation labels for all constellations that have been plotted. This means you must plot the constellations before plotting their labels.
+
+        It's good to plot these last because they're area-based labels (vs point-based, like for star names), and area-based labels have more freedom to move around. If you plot area-based labels first, then it would limit the available space for point-based labels.
+
+        Args:
+            style: Styling of the constellation labels. If None, then the plot's style (specified when creating the plot) will be used
+            label_fn: Callable for determining the label for each constellation. The default function returns the constellation's name in uppercase.
+            collision_handler: An instance of [CollisionHandler][starplot.CollisionHandler] that describes what to do on collisions with other labels, markers, etc. If `None`, then `CollisionHandler(allow_constellation_line_collisions=True)` will be used (**Important: this function does NOT default to the plot's collision handler, since it's the only area-based label function and collisions should be handled differently**).
+        """
+
+        collision_handler = collision_handler or CollisionHandler(
+            allow_constellation_line_collisions=True
+        )
+
         hips = []
         for c in self.objects.constellations:
             hips.extend(c.star_hip_ids)
@@ -283,49 +310,7 @@ class ConstellationPlotterMixin:
                 centroid.x,
                 centroid.y,
                 style,
-                hide_on_collision=self.hide_colliding_labels,
                 area=constellation.boundary,  # TODO : make this intersection with clip path
-                auto_adjust_settings=settings,
+                collision_handler=collision_handler,
                 gid="constellations-label-name",
             )
-
-    def _constellation_labels_static(self, style, label_fn):
-        for constellation in self.objects.constellations:
-            text = label_fn(constellation)
-            self.text(
-                text,
-                constellation.ra,
-                constellation.dec,
-                style,
-                hide_on_collision=self.hide_colliding_labels,
-                remove_on_constellation_collision=False,
-                gid="constellations-label-name",
-            )
-
-    @profile
-    @use_style(LabelStyle, "constellation_labels")
-    def constellation_labels(
-        self,
-        style: LabelStyle = None,
-        label_fn: Callable[[Constellation], str] = Constellation.get_label,
-        auto_adjust: bool = True,
-        auto_adjust_settings: dict = DEFAULT_AUTO_ADJUST_SETTINGS,
-    ):
-        """
-        Plots constellation labels for all constellations that have been plotted. This means you must plot the constellations before plotting their labels.
-
-        It's good to plot these last because they're area-based labels (vs point-based, like for star names), and area-based labels have more freedom to move around. If you plot area-based labels first, then it would limit the available space for point-based labels.
-
-        Args:
-            style: Styling of the constellation labels. If None, then the plot's style (specified when creating the plot) will be used
-            label_fn: Callable for determining the label for each constellation. The default function returns the constellation's name in uppercase.
-            auto_adjust: If True (the default), then labels will be automatically adjusted to avoid collisions with other labels and stars **Important: you must plot stars and constellations first for this to work**. This uses a fairly simple method: for each constellation it finds the centroid of all plotted constellation stars with lines and then generates random points in the constellation boundary starting at the centroid and then progressively increasing the distance from the centroid.
-            auto_adjust_settings: Optional settings for the auto adjustment algorithm.
-        """
-
-        if auto_adjust:
-            settings = DEFAULT_AUTO_ADJUST_SETTINGS
-            settings.update(auto_adjust_settings)
-            self._constellation_labels_auto(style, label_fn, settings=settings)
-        else:
-            self._constellation_labels_static(style, label_fn)