+++ docs/changelog.md
@@ -1,4 +1,13 @@
+## v0.19.x
+- Adds a `CollisionHandler` for defining what to do when labels collide with something
+- Introduces catalogs for the Milky Way and constellation borders
+- Makes map plots default to plotting at epoch J2000
+- Replaces the `true_size` kwarg on the DSO plotting function with `where_true_size` which is a list of filters
+- Fixes a bug with auto label offsets from markers
+
 ## v0.18.x
+[Documentation](https://archives.starplot.dev/0.18.3/)
+
 - Adds function for plotting an arrow, from one point to another or as a pointer to a target
 - Adds support for custom data catalogs for stars, DSOs, and constellations
 - Converts sky object models to data classes, improving performance when many objects are plotted